Use google plus events hangout url for blank archive url
Fix for issue #159.
Use a callback to set a blank archive_url to the hangout_url if the
hangout url is a google plus events url. Set the blank archive url
only when the google plus url contains `events`. Don't set the blank
archive url if the google plus url contains `hoaevent` (even if it
also contains `events`).