Fix lesson display
If a lesson has a hangout url, display "Watch Lesson" only 15 minutes
before the lesson; prior to 15 minutes to the start of the lesson display
"The hangout link will be posted here 15 minutes before the class".

Change when the archive url is displayed, if present, to 5 minutes after
the end of a lesson.