Filter lessons by school
For the `/l/future/slugs` and `/l/upcoming` API endpoints, allow for an
optional :school_id at the end of the url and return only lesson(s) at
the school whose id is passed.
Move the logic for getting the lesson(s), dependent on if a school is
specified, to the lesson model and include tests for the new method used.