Update Calendar Display
Fix for issue #169.
Always update the display of the calendar when there are no more lessons
in the current month and there is at least one lesson in the next month,
even if that lesson is on the 31st of the next month.
Include test that would previously fail for a next lesson on the 31st of
the next month.