Validate Tweet Message Length
Add validation setting a maximum length of 140 for the tweet message on
lesson create and update. Include test and modify controller to handle
a lesson that fails validation.

Check the final tweet message composed in `tweet` method of
`lib/lesson_tweeter.rb`, ensuring that a tweet is not sent if the final
composed tweet message exceeds 140 characters.